fix(ocr): drop the no-op i32 cast in the wasm tesseract backend
TesseractConfig::psm is already i32 (types/formats.rs) and try_from_int takes i32,
so the cast was i32 -> i32 under every feature combination and tripped
clippy::unnecessary_cast.

+++ crates/xberg/src/ocr/tesseract_wasm_backend.rs
@@ -241,7 +241,7 @@ fn resolve_psm(config: &OcrConfig) -> TessPageSegMode {
     config
         .tesseract_config
         .as_ref()
-        .and_then(|c| TessPageSegMode::try_from_int(c.psm as i32))
+        .and_then(|c| TessPageSegMode::try_from_int(c.psm))
         .unwrap_or(DEFAULT_WASM_PSM)
 }
